+++ age/src/plugin.rs
@@ -4,10 +4,11 @@ use age_core::{
     format::{FileKey, Stanza},
     io::{DebugReader, DebugWriter},
     plugin::{Connection, Reply, Response, UnidirSend, IDENTITY_V1, RECIPIENT_V1},
+    primitives::{bech32_decode, bech32_encode},
     secrecy::ExposeSecret,
 };
 use base64::{prelude::BASE64_STANDARD_NO_PAD, Engine};
-use bech32::Variant;
+use bech32::Hrp;
 
 use std::borrow::Borrow;
 use std::collections::HashSet;
@@ -22,14 +23,12 @@ use std::time::{Duration, SystemTime};
 
 use crate::{
     error::{DecryptError, EncryptError, PluginError},
-    fl,
-    util::parse_bech32,
-    wfl, wlnfl, Callbacks,
+    fl, wfl, wlnfl, Callbacks,
 };
 
 // Plugin HRPs are age1[name] and AGE-PLUGIN-[NAME]-
 const PLUGIN_RECIPIENT_PREFIX: &str = "age1";
-const PLUGIN_IDENTITY_PREFIX: &str = "age-plugin-";
+const PLUGIN_IDENTITY_PREFIX: &str = "AGE-PLUGIN-";
 
 pub(crate) const CMD_ERROR: &str = "error";
 const CMD_RECIPIENT_STANZA: &str = "recipient-stanza";
@@ -106,25 +105,31 @@ impl std::str::FromStr for Recipient {
 
     /// Parses a plugin recipient from a string.
     fn from_str(s: &str) -> Result<Self, Self::Err> {
-        parse_bech32(s)
-            .ok_or("invalid Bech32 encoding")
-            .and_then(|(hrp, _)| {
-                if hrp.len() > PLUGIN_RECIPIENT_PREFIX.len()
-                    && hrp.starts_with(PLUGIN_RECIPIENT_PREFIX)
-                {
-                    let name = hrp.split_at(PLUGIN_RECIPIENT_PREFIX.len()).1.to_owned();
-                    if valid_plugin_name(&name) {
-                        Ok(Recipient {
-                            name,
-                            recipient: s.to_owned(),
-                        })
-                    } else {
-                        Err("invalid plugin name")
-                    }
+        bech32_decode(
+            s,
+            |_| "invalid Bech32 encoding",
+            |hrp| {
+                (hrp.len() > PLUGIN_RECIPIENT_PREFIX.len()
+                    && hrp.as_str().starts_with(PLUGIN_RECIPIENT_PREFIX))
+                .then_some(())
+                .ok_or("invalid HRP")
+            },
+            |hrp, _| {
+                let name = hrp
+                    .as_str()
+                    .split_at(PLUGIN_RECIPIENT_PREFIX.len())
+                    .1
+                    .to_owned();
+                if valid_plugin_name(&name) {
+                    Ok(Recipient {
+                        name,
+                        recipient: s.to_owned(),
+                    })
                 } else {
-                    Err("invalid HRP")
+                    Err("invalid plugin name")
                 }
-            })
+            },
+        )
     }
 }
 
@@ -155,30 +160,34 @@ impl std::str::FromStr for Identity {
 
     /// Parses a plugin identity from a string.
     fn from_str(s: &str) -> Result<Self, Self::Err> {
-        parse_bech32(s)
-            .ok_or("invalid Bech32 encoding")
-            .and_then(|(hrp, _)| {
-                if hrp.len() > PLUGIN_IDENTITY_PREFIX.len()
-                    && hrp.starts_with(PLUGIN_IDENTITY_PREFIX)
-                {
-                    // TODO: Decide whether to allow plugin names to end in -
-                    let name = hrp
-                        .split_at(PLUGIN_IDENTITY_PREFIX.len())
-                        .1
-                        .trim_end_matches('-')
-                        .to_owned();
-                    if valid_plugin_name(&name) {
-                        Ok(Identity {
-                            name,
-                            identity: s.to_owned(),
-                        })
-                    } else {
-                        Err("invalid plugin name")
-                    }
+        bech32_decode(
+            s,
+            |_| "invalid Bech32 encoding",
+            |hrp| {
+                (hrp.len() > PLUGIN_IDENTITY_PREFIX.len()
+                    && hrp.as_str().starts_with(PLUGIN_IDENTITY_PREFIX)
+                    && hrp.as_str().ends_with('-'))
+                .then_some(())
+                .ok_or("invalid HRP")
+            },
+            |hrp, _| {
+                // TODO: Decide whether to allow plugin names to end in -
+                let name = hrp
+                    .as_str()
+                    .split_at(PLUGIN_IDENTITY_PREFIX.len())
+                    .1
+                    .trim_end_matches('-')
+                    .to_lowercase();
+                if valid_plugin_name(&name) {
+                    Ok(Identity {
+                        name,
+                        identity: s.to_owned(),
+                    })
                 } else {
-                    Err("invalid HRP")
+                    Err("invalid plugin name")
                 }
-            })
+            },
+        )
     }
 }
 
@@ -190,24 +199,19 @@ impl fmt::Display for Identity {
 
 impl Identity {
     /// Returns the identity corresponding to the given plugin name in its default mode.
-    ///
-    /// # Panics
-    ///
-    /// Panics if `plugin_name` contains invalid characters.
-    pub fn default_for_plugin(plugin_name: &str) -> Self {
+    pub fn default_for_plugin(plugin_name: &str) -> Result<Self, ResolveError> {
         if valid_plugin_name(plugin_name) {
-            bech32::encode(
-                &format!("{}{}-", PLUGIN_IDENTITY_PREFIX, plugin_name),
-                [],
-                Variant::Bech32,
+            Ok(bech32_encode(
+                Hrp::parse_unchecked(&format!("{}{}-", PLUGIN_IDENTITY_PREFIX, plugin_name)),
+                &[],
             )
-            .expect("HRP is valid")
             .to_uppercase()
             .parse()
-            .unwrap()
+            .unwrap())
         } else {
-            // TODO: Change the API to be fallible.
-            panic!("invalid plugin name")
+            Err(ResolveError::InvalidPluginName {
+                name: plugin_name.into(),
+            })
         }
     }
 
@@ -225,9 +229,11 @@ pub(crate) struct Plugin {
 
 impl Plugin {
     /// Finds the age plugin with the given name in `$PATH`.
-    ///
-    /// On error, returns the binary name that could not be located.
-    pub(crate) fn new(name: &str) -> Result<Self, String> {
+    pub(crate) fn new(name: &str) -> Result<Self, ResolveError> {
+        if !valid_plugin_name(name) {
+            return Err(ResolveError::InvalidPluginName { name: name.into() });
+        }
+
         let binary_name = binary_name(name);
         match which::which(&binary_name).or_else(|e| {
             // If we are running in WSL, try appending `.exe`; plugins installed in
@@ -240,7 +246,7 @@ impl Plugin {
             }
         }) {
             Ok(path) => Ok(Plugin { binary_name, path }),
-            Err(_) => Err(binary_name),
+            Err(_) => Err(ResolveError::MissingPlugin { binary_name }),
         }
     }
 
@@ -386,29 +392,21 @@ impl<C: Callbacks> RecipientPluginV1<C> {
         recipients: &[Recipient],
         identities: &[Identity],
         callbacks: C,
-    ) -> Result<Self, EncryptError> {
-        if valid_plugin_name(plugin_name) {
-            Plugin::new(plugin_name)
-                .map_err(|binary_name| EncryptError::MissingPlugin { binary_name })
-                .map(|plugin| RecipientPluginV1 {
-                    plugin,
-                    recipients: recipients
-                        .iter()
-                        .filter(|r| r.name == plugin_name)
-                        .cloned()
-                        .collect(),
-                    identities: identities
-                        .iter()
-                        .filter(|r| r.name == plugin_name)
-                        .cloned()
-                        .collect(),
-                    callbacks,
-                })
-        } else {
-            Err(EncryptError::MissingPlugin {
-                binary_name: plugin_name.to_string(),
-            })
-        }
+    ) -> Result<Self, ResolveError> {
+        Plugin::new(plugin_name).map(|plugin| RecipientPluginV1 {
+            plugin,
+            recipients: recipients
+                .iter()
+                .filter(|r| r.name == plugin_name)
+                .cloned()
+                .collect(),
+            identities: identities
+                .iter()
+                .filter(|r| r.name == plugin_name)
+                .cloned()
+                .collect(),
+            callbacks,
+        })
     }
 }
 
@@ -610,23 +608,15 @@ impl<C: Callbacks> IdentityPluginV1<C> {
         plugin_name: &str,
         identities: &[Identity],
         callbacks: C,
-    ) -> Result<Self, DecryptError> {
-        if valid_plugin_name(plugin_name) {
-            Plugin::new(plugin_name)
-                .map_err(|binary_name| DecryptError::MissingPlugin { binary_name })
-                .map(|plugin| {
-                    let identities = identities
-                        .iter()
-                        .filter(|r| r.name == plugin_name)
-                        .cloned()
-                        .collect();
-                    Self::from_parts(plugin, identities, callbacks)
-                })
-        } else {
-            Err(DecryptError::MissingPlugin {
-                binary_name: plugin_name.to_string(),
-            })
-        }
+    ) -> Result<Self, ResolveError> {
+        Plugin::new(plugin_name).map(|plugin| {
+            let identities = identities
+                .iter()
+                .filter(|r| r.name == plugin_name)
+                .cloned()
+                .collect();
+            Self::from_parts(plugin, identities, callbacks)
+        })
     }
 
     pub(crate) fn from_parts(plugin: Plugin, identities: Vec<Identity>, callbacks: C) -> Self {
@@ -758,13 +748,48 @@ impl<C: Callbacks> crate::Identity for IdentityPluginV1<C> {
     }
 }
 
+/// Errors returned when resolving a plugin.
+#[derive(Clone, Debug)]
+#[non_exhaustive]
+pub enum ResolveError {
+    /// A provided plugin name was invalid.
+    InvalidPluginName {
+        /// The provided name.
+        name: String,
+    },
+    /// A required plugin could not be found.
+    MissingPlugin {
+        /// The plugin's binary name.
+        binary_name: String,
+    },
+}
+
+impl fmt::Display for ResolveError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            ResolveError::InvalidPluginName { name } => {
+                wfl!(f, "err-invalid-plugin-name", name = name.as_str())
+            }
+            ResolveError::MissingPlugin { binary_name } => {
+                wlnfl!(f, "err-missing-plugin", plugin_name = binary_name.as_str())?;
+                wfl!(f, "rec-missing-plugin")
+            }
+        }
+    }
+}
+
+impl std::error::Error for ResolveError {}
+
 #[cfg(test)]
 mod tests {
-    use crate::{DecryptError, EncryptError, NoCallbacks};
+    use age_core::primitives::bech32_encode;
+    use bech32::Hrp;
+
+    use crate::NoCallbacks;
 
     use super::{
-        Identity, IdentityPluginV1, Recipient, RecipientPluginV1, PLUGIN_IDENTITY_PREFIX,
-        PLUGIN_RECIPIENT_PREFIX,
+        Identity, IdentityPluginV1, Recipient, RecipientPluginV1, ResolveError,
+        PLUGIN_IDENTITY_PREFIX, PLUGIN_RECIPIENT_PREFIX,
     };
 
     const INVALID_PLUGIN_NAME: &str = "foobar/../../../../../../../usr/bin/echo";
@@ -772,70 +797,73 @@ mod tests {
     #[test]
     fn default_for_plugin() {
         assert_eq!(
-            Identity::default_for_plugin("foobar").to_string(),
+            Identity::default_for_plugin("foobar").unwrap().to_string(),
             "AGE-PLUGIN-FOOBAR-1QVHULF",
         );
     }
 
     #[test]
     fn recipient_rejects_empty_name() {
-        let invalid_recipient =
-            bech32::encode(PLUGIN_RECIPIENT_PREFIX, [], bech32::Variant::Bech32).unwrap();
+        let invalid_recipient = bech32_encode(Hrp::parse_unchecked(PLUGIN_RECIPIENT_PREFIX), &[]);
         assert!(invalid_recipient.parse::<Recipient>().is_err());
     }
 
     #[test]
     fn recipient_rejects_invalid_chars() {
-        let invalid_recipient = bech32::encode(
-            &format!("{}{}", PLUGIN_RECIPIENT_PREFIX, INVALID_PLUGIN_NAME),
-            [],
-            bech32::Variant::Bech32,
-        )
-        .unwrap();
+        let invalid_recipient = bech32_encode(
+            Hrp::parse_unchecked(&format!(
+                "{}{}",
+                PLUGIN_RECIPIENT_PREFIX, INVALID_PLUGIN_NAME
+            )),
+            &[],
+        );
         assert!(invalid_recipient.parse::<Recipient>().is_err());
     }
 
     #[test]
     fn identity_rejects_empty_name() {
-        let invalid_identity = bech32::encode(
-            &format!("{}-", PLUGIN_IDENTITY_PREFIX),
-            [],
-            bech32::Variant::Bech32,
+        let invalid_identity = bech32_encode(
+            Hrp::parse_unchecked(&format!("{}-", PLUGIN_IDENTITY_PREFIX)),
+            &[],
         )
-        .expect("HRP is valid")
         .to_uppercase();
         assert!(invalid_identity.parse::<Identity>().is_err());
     }
 
     #[test]
     fn identity_rejects_invalid_chars() {
-        let invalid_identity = bech32::encode(
-            &format!("{}{}-", PLUGIN_IDENTITY_PREFIX, INVALID_PLUGIN_NAME),
-            [],
-            bech32::Variant::Bech32,
+        let invalid_identity = bech32_encode(
+            Hrp::parse_unchecked(&format!(
+                "{}{}-",
+                PLUGIN_IDENTITY_PREFIX, INVALID_PLUGIN_NAME
+            )),
+            &[],
         )
-        .expect("HRP is valid")
         .to_uppercase();
         assert!(invalid_identity.parse::<Identity>().is_err());
     }
 
     #[test]
-    #[should_panic]
     fn identity_default_for_plugin_rejects_empty_name() {
-        Identity::default_for_plugin("");
+        assert!(matches!(
+            Identity::default_for_plugin(""),
+            Err(ResolveError::InvalidPluginName { name }) if name.is_empty(),
+        ));
     }
 
     #[test]
-    #[should_panic]
     fn identity_default_for_plugin_rejects_invalid_chars() {
-        Identity::default_for_plugin(INVALID_PLUGIN_NAME);
+        assert!(matches!(
+            Identity::default_for_plugin(INVALID_PLUGIN_NAME),
+            Err(ResolveError::InvalidPluginName { name }) if name == INVALID_PLUGIN_NAME,
+        ));
     }
 
     #[test]
     fn recipient_plugin_v1_rejects_empty_name() {
         assert!(matches!(
             RecipientPluginV1::new("", &[], &[], NoCallbacks),
-            Err(EncryptError::MissingPlugin { binary_name }) if binary_name.is_empty(),
+            Err(ResolveError::InvalidPluginName { name }) if name.is_empty(),
         ));
     }
 
@@ -843,7 +871,7 @@ mod tests {
     fn recipient_plugin_v1_rejects_invalid_chars() {
         assert!(matches!(
             RecipientPluginV1::new(INVALID_PLUGIN_NAME, &[], &[], NoCallbacks),
-            Err(EncryptError::MissingPlugin { binary_name }) if binary_name == INVALID_PLUGIN_NAME,
+            Err(ResolveError::InvalidPluginName { name }) if name == INVALID_PLUGIN_NAME,
         ));
     }
 
@@ -851,7 +879,7 @@ mod tests {
     fn identity_plugin_v1_rejects_empty_name() {
         assert!(matches!(
             IdentityPluginV1::new("", &[], NoCallbacks),
-            Err(DecryptError::MissingPlugin { binary_name }) if binary_name.is_empty(),
+            Err(ResolveError::InvalidPluginName { name }) if name.is_empty(),
         ));
     }
 
@@ -859,7 +887,7 @@ mod tests {
     fn identity_plugin_v1_rejects_invalid_chars() {
         assert!(matches!(
             IdentityPluginV1::new(INVALID_PLUGIN_NAME, &[], NoCallbacks),
-            Err(DecryptError::MissingPlugin { binary_name }) if binary_name == INVALID_PLUGIN_NAME,
+            Err(ResolveError::InvalidPluginName { name }) if name == INVALID_PLUGIN_NAME,
         ));
     }
 }