+++ rage/src/bin/rage-mount/zip.rs
@@ -26,7 +26,7 @@ fn zipfile_to_fuse(zf: &ZipFile) -> FileAttr {
     let mtime: SystemTime = zf.last_modified().to_time().unwrap().into();
 
     FileAttr {
-        size: zf.size() as u64,
+        size: zf.size(),
         blocks: 1,
         atime: mtime,
         mtime,
@@ -107,7 +107,7 @@ impl AgeZipFs {
                 .by_index(i)
                 .map_err(|e| io::Error::new(io::ErrorKind::Other, e))?;
             if let Some(path) = zf.enclosed_name() {
-                add_dir_to_map(&mut dir_map, path, zipfile_to_filetype(&zf));
+                add_dir_to_map(&mut dir_map, &path, zipfile_to_filetype(&zf));
             }
         }
 
@@ -202,7 +202,7 @@ impl FilesystemMT for AgeZipFs {
         let mut open_files = self.open_files.lock().unwrap();
 
         for i in 0..inner.len() {
-            if inner.by_index(i).unwrap().enclosed_name() == Some(zip_path(path)) {
+            if inner.by_index(i).unwrap().enclosed_name().as_deref() == Some(zip_path(path)) {
                 let fh = open_files.1;
                 open_files.0.insert(fh, i);
                 open_files.1 = open_files.1.wrapping_add(1);
@@ -233,8 +233,7 @@ impl FilesystemMT for AgeZipFs {
                 }
 
                 // Skip to offset
-                let mut buf = vec![];
-                buf.resize(offset as usize, 0);
+                let mut buf = vec![0; offset as usize];
                 if zf.read_exact(&mut buf).is_err() {
                     return callback(Err(libc::EIO));
                 }