+++ age/CHANGELOG.md
@@ -9,6 +9,36 @@ and this project adheres to Rust's notion of
 to 1.0.0 are beta releases.
 
 ## [Unreleased]
+### Added
+- Support for new native age recipient types:
+  - `age::tag::Recipient` (encryption-only)
+  - `age::tagpq::Recipient` (encryption-only)
+- `age::encrypted::EncryptedIdentity`
+- `age::plugin::ResolveError`
+
+### Changed
+- MSRV is now 1.74.0.
+- Migrated to `base64 0.22`, `i18n-embed 0.16`.
+- `age::IdentityFile::into_identities` now returns
+  `Result<Vec<Box<dyn crate::Identity + Send + Sync>>, DecryptError>` instead of
+  `Result<Vec<Box<dyn crate::Identity>>, DecryptError>`. This re-enables
+  cross-thread uses of `IdentityFile`, which were unintentionally disabled in
+  0.11.0.
+- `age::plugin`:
+  - The following methods now returns `Result<Self, ResolveError>`:
+    - `Identity::default_for_plugin`
+    - `RecipientPluginV1::new`
+    - `IdentityPluginV1::new`
+- All existing error enums nameable in the public API are now non-exhaustive:
+  - `age::{EncryptError, DecryptError}`
+  - `age::IdentityFileConvertError`
+  - `age::armor::ArmoredReadError`
+  - `age::cli_common::ReadError`
+  - `age::ssh::ParseRecipientKeyError`
+- Removed the following error enum variants:
+  - `age::DecryptError::MissingPlugin`
+  - `age::EncryptError::MissingPlugin`
+  - `age::cli_common::ReadError::MissingPlugin`
 
 ## [0.11.3] - 2026-04-22
 ### Changed